# api.yaml
swagger: "2.0"
info:
  description: "df-gitter api"
  version: "1.0.0"
  title: "Datafoundry Gitter Api"
  #termsOfService: "http://swagger.io/terms/"
  #contact:
    #email: "apiteam@swagger.io"
  #license:
    #name: "Apache 2.0"
    #url: "http://www.apache.org/licenses/LICENSE-2.0.html"
#host: "petstore.swagger.io"
basePath: "/v1"
...
schemes:
- "https"
- "http"
paths:
  /authorize/{source}:
    get:
      tags:
      - "authorize"
      summary: "查询授权"
      description: ""
      operationId: "authorize"
      produces:
      - "application/json"
      parameters:
      - name: "source"
        in: "path"
        description: "gitlab或者github授权"
        required: true
        type: "string"
      responses:
        400:
          description: "跳转授权"
          schema:
            $ref: "#/definitions/Error"
...
definitions:
  Error:
    type: "object"
    properties:
      code:
        type: "integer"
        format: "int32"
      message:
        type: "string"
...